Added unit tests for Asp.Net MVC apps
This adds a Test assets project that is an Asp.Net MVC app with WebApi 2.0 and a new unit test project that tests against it.

This is not working yet, but I wanted to capture the framework for it. There are still some errors of missing DLLs when roslyn opens the project.